+++ crates/adapters/databento/src/python/loader.rs
@@ -133,6 +133,7 @@ impl DatabentoDataLoader {
                 instrument_id,
                 price_precision,
                 include_trades.unwrap_or(false),
+                None,
             )
             .map_err(to_pyvalue_err)?;
 
@@ -161,7 +162,7 @@ impl DatabentoDataLoader {
         price_precision: Option<u8>,
     ) -> PyResult<PyObject> {
         let iter = self
-            .read_records::<dbn::Mbp10Msg>(&filepath, instrument_id, price_precision, false)
+            .read_records::<dbn::Mbp10Msg>(&filepath, instrument_id, price_precision, false, None)
             .map_err(to_pyvalue_err)?;
 
         exhaust_data_iter_to_pycapsule(py, iter).map_err(to_pyvalue_err)
@@ -195,6 +196,7 @@ impl DatabentoDataLoader {
                 instrument_id,
                 price_precision,
                 include_trades.unwrap_or(false),
+                None,
             )
             .map_err(to_pyvalue_err)?;
 
@@ -223,7 +225,7 @@ impl DatabentoDataLoader {
         price_precision: Option<u8>,
     ) -> PyResult<PyObject> {
         let iter = self
-            .read_records::<dbn::BboMsg>(&filepath, instrument_id, price_precision, false)
+            .read_records::<dbn::BboMsg>(&filepath, instrument_id, price_precision, false, None)
             .map_err(to_pyvalue_err)?;
 
         exhaust_data_iter_to_pycapsule(py, iter).map_err(to_pyvalue_err)
@@ -251,7 +253,7 @@ impl DatabentoDataLoader {
         price_precision: Option<u8>,
     ) -> PyResult<PyObject> {
         let iter = self
-            .read_records::<dbn::TbboMsg>(&filepath, instrument_id, price_precision, false)
+            .read_records::<dbn::TbboMsg>(&filepath, instrument_id, price_precision, false, None)
             .map_err(to_pyvalue_err)?;
 
         exhaust_data_iter_to_pycapsule(py, iter).map_err(to_pyvalue_err)
@@ -279,35 +281,48 @@ impl DatabentoDataLoader {
         price_precision: Option<u8>,
     ) -> PyResult<PyObject> {
         let iter = self
-            .read_records::<dbn::TradeMsg>(&filepath, instrument_id, price_precision, false)
+            .read_records::<dbn::TradeMsg>(&filepath, instrument_id, price_precision, false, None)
             .map_err(to_pyvalue_err)?;
 
         exhaust_data_iter_to_pycapsule(py, iter).map_err(to_pyvalue_err)
     }
 
     #[pyo3(name = "load_bars")]
-    #[pyo3(signature = (filepath, instrument_id=None, price_precision=None))]
+    #[pyo3(signature = (filepath, instrument_id=None, price_precision=None, timestamp_on_close=true))]
     fn py_load_bars(
         &self,
         filepath: PathBuf,
         instrument_id: Option<InstrumentId>,
         price_precision: Option<u8>,
+        timestamp_on_close: bool,
     ) -> PyResult<Vec<Bar>> {
-        self.load_bars(&filepath, instrument_id, price_precision)
-            .map_err(to_pyvalue_err)
+        self.load_bars(
+            &filepath,
+            instrument_id,
+            price_precision,
+            Some(timestamp_on_close),
+        )
+        .map_err(to_pyvalue_err)
     }
 
     #[pyo3(name = "load_bars_as_pycapsule")]
-    #[pyo3(signature = (filepath, instrument_id=None, price_precision=None))]
+    #[pyo3(signature = (filepath, instrument_id=None, price_precision=None, timestamp_on_close=true))]
     fn py_load_bars_as_pycapsule(
         &self,
         py: Python,
         filepath: PathBuf,
         instrument_id: Option<InstrumentId>,
         price_precision: Option<u8>,
+        timestamp_on_close: bool,
     ) -> PyResult<PyObject> {
         let iter = self
-            .read_records::<dbn::OhlcvMsg>(&filepath, instrument_id, price_precision, false)
+            .read_records::<dbn::OhlcvMsg>(
+                &filepath,
+                instrument_id,
+                price_precision,
+                false,
+                Some(timestamp_on_close),
+            )
             .map_err(to_pyvalue_err)?;
 
         exhaust_data_iter_to_pycapsule(py, iter).map_err(to_pyvalue_err)